refactored cache refresh
Only store a response in the cache if the status code is < 300. Also
save the response body early on so that it can be read for cache
purposes, but also sent straight back to the requester, and don't send
the response back from the CacheData object that was prepared for
storing in the cache.