Add caching support in the local proxy server (#58)
Adds caching support to the local proxy server for the Vault lambda
extension. To use, set the VAULT_DEFAULT_CACHE_TTL environment
variable to a time.Duration value (15m, 1h). Requests with the header
"X-Vault-Cache-Control: cache" will be returned from the cache on a
hit, and responses will be stored in the cache if not. The cache can
be invalidated by setting the header to "recache".

Also if the VAULT_DEFAULT_CACHE_ENABLE environment variable is set to
true, all requests will behave as though "X-Vault-Cache-Control:
cache" was set. Setting the header to `nocache` on a request will
opt-out of caching entirely in this configuration.

Co-authored-by: Sunny Wang <wangjinlian@hotmail.com>
Co-authored-by: Jinlian(Sunny) Wang <JinlianWang@users.noreply.github.com>
Co-authored-by: Tom Proctor <tomhjp@users.noreply.github.com>