fix arm64 build and version metadata
This change ensures we build for both arm64 and amd64 as defined in the
build matrix.

In https://github.com/hashicorp/vault-lambda-extension/pull/155, we
removed some environment variables in the build workflow file that were
required to properly build the executable for all systems defined in the
build matrix.

This caused the `make zip` command to always rely on the default values
defined in the makefile, causing problems with the used goarch and
version (GOOS?=linux, GOARCH?=amd64, VERSION?=0.0.0-dev).

Closes https://github.com/hashicorp/vault-lambda-extension/issues/174